separação do chat e da main
